refactor: Customer 패키지 이동
