+++ src/main/java/woowacourse/auth/application/AuthService.java
@@ -6,6 +6,7 @@ import woowacourse.auth.dto.token.TokenRequest;
 import woowacourse.auth.dto.token.TokenResponse;
 import woowacourse.auth.exception.InvalidAuthException;
 import woowacourse.auth.support.JwtTokenProvider;
+import woowacourse.shoppingcart.application.CustomerService;
 
 @Service
 public class AuthService {
