+++ src/main/java/woowacourse/auth/ui/config/AuthenticationPrincipalArgumentResolver.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import woowacourse.auth.application.CustomerService;
+import woowacourse.shoppingcart.application.CustomerService;
 import woowacourse.auth.domain.Customer;
 import woowacourse.auth.support.AuthorizationExtractor;
 import woowacourse.auth.support.JwtTokenProvider;
